LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/svg-script-xlinkhref.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <svg><script xlink:href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=58afcb59-649a-4d57-934e-594044c51cf9&amp;encodingcheck=&Gbreve;"></script></svg>


FAIL Speculative parsing, document.write(): svg-script-xlinkhref Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
